Moving remove_squeezable_dimensions to framework/tensor_util so that it can be used outside metrics. Change: 134449221
