+++ tensorflow/contrib/framework/python/framework/tensor_util_test.py
@@ -296,5 +296,108 @@ class ConvertToTensorOrSparseTensorTest(tf.test.TestCase):
         self.assertAllEqual(sparse_tensor_value.shape, convertee.shape)
 
 
+class RemoveSqueezableDimensionsTest(tf.test.TestCase):
+
+  def testRemoveSqueezableDimensions(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=False, predictions_have_extra_dim=False,
+        labels_have_static_shape=False, labels_have_extra_dim=False)
+
+  def testRemoveSqueezableDimensions_extraLabelDim(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=False, predictions_have_extra_dim=False,
+        labels_have_static_shape=False, labels_have_extra_dim=True)
+
+  def testRemoveSqueezableDimensions_staticLabel(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=False, predictions_have_extra_dim=False,
+        labels_have_static_shape=True, labels_have_extra_dim=False)
+
+  def testRemoveSqueezableDimensions_staticLabel_extraLabelDim(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=False, predictions_have_extra_dim=False,
+        labels_have_static_shape=True, labels_have_extra_dim=True)
+
+  def testRemoveSqueezableDimensions_extraPredictionDim(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=False, predictions_have_extra_dim=True,
+        labels_have_static_shape=False, labels_have_extra_dim=False)
+
+  def testRemoveSqueezableDimensions_extraPredictionDim_staticLabel(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=False, predictions_have_extra_dim=True,
+        labels_have_static_shape=True, labels_have_extra_dim=False)
+
+  def testRemoveSqueezableDimensions_staticPrediction(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=True, predictions_have_extra_dim=False,
+        labels_have_static_shape=False, labels_have_extra_dim=False)
+
+  def testRemoveSqueezableDimensions_staticPrediction_extraLabelDim(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=True, predictions_have_extra_dim=False,
+        labels_have_static_shape=False, labels_have_extra_dim=True)
+
+  def testRemoveSqueezableDimensions_static(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=True, predictions_have_extra_dim=False,
+        labels_have_static_shape=True, labels_have_extra_dim=False)
+
+  def testRemoveSqueezableDimensions_static_extraLabelDim(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=True, predictions_have_extra_dim=False,
+        labels_have_static_shape=True, labels_have_extra_dim=True)
+
+  def testRemoveSqueezableDimensions_staticPrediction_extraPredictionDim(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=True, predictions_have_extra_dim=True,
+        labels_have_static_shape=False, labels_have_extra_dim=False)
+
+  def testRemoveSqueezableDimensions_static_extraPredictionDim(self):
+    self._testRemoveSqueezableDimensions(
+        predictions_have_static_shape=True, predictions_have_extra_dim=True,
+        labels_have_static_shape=True, labels_have_extra_dim=False)
+
+  # TODO(ptucker): Replace this with parameterized test.
+  def _testRemoveSqueezableDimensions(
+      self,
+      predictions_have_static_shape,
+      predictions_have_extra_dim,
+      labels_have_static_shape,
+      labels_have_extra_dim):
+    assert not (predictions_have_extra_dim and labels_have_extra_dim)
+    predictions_value = (0, 1, 1, 0, 0, 1, 0)
+    labels_value = (0, 0, 1, 1, 0, 0, 0)
+
+    input_predictions_value = (
+        [[p] for p in predictions_value] if predictions_have_extra_dim else
+        predictions_value)
+    input_labels_value = (
+        [[l] for l in labels_value] if labels_have_extra_dim else labels_value)
+
+    with tf.Graph().as_default() as g:
+      feed_dict = {}
+      if predictions_have_static_shape:
+        predictions = tf.constant(input_predictions_value, dtype=tf.int32)
+      else:
+        predictions = tf.placeholder(dtype=tf.int32, name="predictions")
+        feed_dict[predictions] = input_predictions_value
+      if labels_have_static_shape:
+        labels = tf.constant(input_labels_value, dtype=tf.int32)
+      else:
+        labels = tf.placeholder(dtype=tf.int32, name="labels")
+        feed_dict[labels] = input_labels_value
+
+      squeezed_predictions, squeezed_labels = (
+          tf.contrib.framework.remove_squeezable_dimensions(
+              predictions, labels))
+      with self.test_session(g):
+        tf.initialize_local_variables().run()
+        self.assertAllClose(
+            predictions_value, squeezed_predictions.eval(feed_dict=feed_dict))
+        self.assertAllClose(
+            labels_value, squeezed_labels.eval(feed_dict=feed_dict))
+
+
 if __name__ == "__main__":
   tf.test.main()
