+++ tensorflow/contrib/metrics/BUILD
@@ -137,18 +137,6 @@ py_test(
 )
 
 py_test(
-    name = "metric_ops_util_test",
-    srcs = ["python/ops/metric_ops_util_test.py"],
-    srcs_version = "PY2AND3",
-    deps = [
-        ":metrics_py",
-        "//tensorflow:tensorflow_py",
-        "//tensorflow/python:framework_test_lib",
-        "//tensorflow/python:platform_test",
-    ],
-)
-
-py_test(
     name = "metric_ops_test",
     srcs = ["python/ops/metric_ops_test.py"],
     shard_count = 3,
